changes to loggeddashboardnumbers to loggednetworknumbers
